Selenium/Page_object_model/Autotests_online_pizzeria/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver==2.24.1
chromedriver-binary==124.0.6342.2.0
chromedriver-py==121.0.6167.85
colorama==0.4.6
cryptography==42.0.5
flake8==7.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
kaitaistruct==0.10
mccabe==0.7.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydivert==2.1.0
pyflakes==3.2.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
zstandard==0.22.0
